Used getStaticProps to get data for our static page before the rendering of the page.
The reason we use a custom Next feature is that before we use getStaticProps we had to use useState and useEffect to send HTTP request to our backends and get data, but in that process useEffect doesn't renders when the page first rendered and therefore we get an empty HTML source file and it is bad for SEO to not have data. This is why we use getStaticProps to render the data on the deployment stage too.